library: Makefile: rework build rules
There's a bit of duplication between the build rules under the `lib` and
`clean` build rule. So all build targets are under the TARGETS variable
now, and the rules are generated by iterating that var.

Also added `+` prefix to $(MAKE). This should essentially parallelize the
build when calling it with `-jN`.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>