adrv900zu11eg: Data Offload integration
Tx and Rx path have a DO controller with and external memory for
buffering.

Note: at this point there are no external syncronization connected to
the DO controller. The positive edge of the xfer_req will trigger the
capture, in case of the ADC, and the end of the initialisation will
trigger the transmittion in case of DAC.